requirements.txt
opencv-python~=4.4.0.46
numpy~=1.19.4
Pillow~=8.0.1
dnspython~=2.0.0
ping3~=2.6.6
requests~=2.22.0
selenium~=3.141.0
Flask~=1.1.2
beautifulsoup4~=4.9.3
googletrans~=3.0.0
uuid~=1.30
WTForms~=2.2.1
Jinja2~=2.11.1
sendgrid~=5.3.0
click~=7.1.1
redis~=3.4.1
jieba~=0.42.1
PyMySQL~=0.9.3
zhconv~=1.4.1
xlwt~=1.3.0
tinify~=1.5.1
openpyxl~=3.0.5